🛡️ Sentinel: [security improvement] Harden unique ID generation (fix build)
This PR centralizes and hardens the generation of unique identifiers across the application, with fixes for the Cloudflare build.

Key changes:
- Introduced a cryptographically secure `generateSecureId` utility in `src/lib/utils.ts` that uses `crypto.randomUUID()` with secure fallbacks.
- Refactored multiple modules (session analytics, A/B testing, event bus, etc.) to use the new secure utility.
- Replaced an insecure `console.log` in `src/lib/session-analytics.ts` with a proper logger call.
- Hardened ID generation in `src/lib/config/timeline.ts` with a local secure implementation to avoid circular dependencies and build failures in the config layer.

Verified with `pnpm security:check` and a successful `pnpm run build:cloudflare`.
